[LS-030] Refactored Database Tests  [BJ, RT]
