+++ app/src/androidTest/java/com/example/loginsesame/database/TestDatabaseVaultEntries.kt
@@ -1,4 +1,4 @@
-package com.example.loginsesame
+package com.example.loginsesame.database
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
@@ -12,14 +12,14 @@ import org.junit.Test
 import java.io.IOException
 import kotlin.jvm.Throws
 
-class VaultEntryTable {
+class TestDatabaseVaultEntries {
 
     private lateinit var vaultEntryDao: VaultEntryDao
     private lateinit var userDao: UserDao
     private lateinit var db: UserDatabase
 
     @Before
-    fun createDb() {
+    fun initDb() {
         val context = ApplicationProvider.getApplicationContext<Context>()
         db = UserDatabase.initDb(context)
         userDao = db.getUserDao()
@@ -32,7 +32,7 @@ class VaultEntryTable {
 
     @After
     @Throws(IOException::class)
-    fun closeDb() {
+    fun cleanup() {
         userDao.deleteAllUsers()
         vaultEntryDao.deleteAllEntries()
         //db.close()
@@ -40,25 +40,25 @@ class VaultEntryTable {
 
 
     @Test
-    fun addEntity(){
+    fun testAddEntity(){
         val entity = VaultEntry(2, "account_z", "user_z", "password")
         vaultEntryDao.add(entity)
         assert(vaultEntryDao.allEntries().get(1) == entity)
     }
 
     @Test
-    fun vaultEntry(){
+    fun testGetEntity(){
         assert(vaultEntryDao.getEntity("account_x") == vaultEntryDao.allEntries().get(0))
     }
 
     @Test
-    fun removeEntity(){
+    fun testDeleteEntity(){
         vaultEntryDao.deleteVaultEntry(vaultEntryDao.allEntries().get(0))
         assert(vaultEntryDao.getEntity("account_x") == null)
     }
 
     @Test
-    fun editEntity(){
+    fun testEditEntity(){
         val entity_b = VaultEntry(3, "account_b", "user_b", "password")
         vaultEntryDao.add(entity_b)
         val entity = VaultEntry(3, "account_y", "user_y", "password_y")
